fix allocator

+++ include/stdex/pool.h
@@ -8,7 +8,7 @@
 
 namespace stdex
 {
-    template<class TBlockType, size_t TBlockCount>
+    template<class TBlockType, uint32_t TBlockCount>
     struct pool_template_chunk
     {
         typedef pool_template_chunk<TBlockType, TBlockCount> chunk_t;
@@ -71,7 +71,7 @@ namespace stdex
 		}
 	};
 
-	template<class TBlockType, size_t TBlockCount, class TAllocator = stdex_pool_allocator_default>
+	template<class TBlockType, uint32_t TBlockCount, class TAllocator = stdex_pool_allocator_default>
     class pool
     {
         typedef pool_template_chunk<TBlockType, TBlockCount> chunk_t;
@@ -125,7 +125,7 @@ namespace stdex
         }
 
 	public:
-        size_t getBlockCount() const
+        uint32_t getBlockCount() const
         {
             return m_blockCount;
         }
@@ -135,12 +135,12 @@ namespace stdex
             return sizeof( TBlockType );
         }
 
-        size_t getChunkCount() const
+        uint32_t getChunkCount() const
         {
             return m_chunkCount;
         }
 
-        size_t getBlockTotal() const
+        uint32_t getBlockTotal() const
         {
             return TBlockCount;
         }
@@ -210,11 +210,11 @@ namespace stdex
         chunk_t * m_chunk;
         block_t * m_free;
 
-        size_t m_blockCount;
-        size_t m_chunkCount;
+        uint32_t m_blockCount;
+        uint32_t m_chunkCount;
     };
 
-    template<class TBlockType, size_t TBlockCount>
+    template<class TBlockType, uint32_t TBlockCount>
     class template_pool
     {
     public:
